requirements.txt
requests==2.21.0
schedule==0.6.0
GitPython==3.1.12
